[Fix] 시즌 드롭다운 리코일 관리 #375
