+++ components/mode/ModeSeasonMineProvider.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import axios from 'axios';
 import { Mode } from 'types/mainType';
-import { Seasons } from 'types/seasonTypes';
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
+import { useRecoilValue } from 'recoil';
+import { seasonState } from 'utils/recoil/seasons';
 import ModeToggle from './ModeToggle';
 import SeasonDropDown from './SeasonDropDown';
 import ModeRadiobox from './ModeRadiobox';
@@ -16,21 +16,7 @@ export default function ModeSeasonProvider({ children }: ModeSelectProps) {
   const [mode, setMode] = useState<Mode>('both');
   const [isMine, setIsMine] = useState(false);
   const [season, setSeason] = useState('');
-  const [seasonList, setSeasonList] = useState<Seasons>();
-
-  useEffect(() => {
-    getSeasonListHandler();
-  }, [mode]);
-
-  const getSeasonListHandler = async () => {
-    try {
-      // const res = await instance.get(`/pingpong/seasonlist`);
-      const res = await axios.get(
-        `http://localhost:3000/api/pingpong/seasonlist`
-      ); // api 연결 후 삭제 예정
-      setSeasonList(res?.data);
-    } catch (e) {}
-  };
+  const seasonList = useRecoilValue(seasonState);
 
   const modeChangeHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
     setMode(e.target.value as Mode);
